separated into main.dart and detailPage.dart
