+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:helloworld/detailPage.dart';
 
 void main() {
   runApp(MyApp());
@@ -42,20 +43,5 @@ class ListPage extends StatelessWidget {
           itemCount: 10,
       ),
      );
-  }
-}
-
-class DetailPage extends StatelessWidget{
-  DetailPage(this.index);
-  final int index;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        child: Center(
-          child: Text('$index', style: TextStyle(fontSize:28.0),),
-        )
-      ),
-    );
   }
 }
